refactor(admin): move KMS management handlers
